Push to develop Sprint III v1

+++ Development/ChatServer/src/main/java/edu/northeastern/ccs/im/service/jpa_service/GroupJPAService.java
@@ -1,9 +1,15 @@
 package edu.northeastern.ccs.im.service.jpa_service;
 
 import java.util.List;
+import java.util.logging.Logger;
 
 import javax.persistence.*;
 
+import edu.northeastern.ccs.im.ChatLogger;
+import edu.northeastern.ccs.im.customexceptions.GroupNotDeletedException;
+import edu.northeastern.ccs.im.customexceptions.GroupNotFoundException;
+import edu.northeastern.ccs.im.customexceptions.GroupNotPersistedException;
+import edu.northeastern.ccs.im.customexceptions.UserNotFoundException;
 import edu.northeastern.ccs.im.user_group.Group;
 import edu.northeastern.ccs.im.user_group.User;
 
@@ -12,7 +18,10 @@ import edu.northeastern.ccs.im.user_group.User;
  */
 public class GroupJPAService{
 
-	/** The emfactory. */
+	private static final Logger LOGGER = Logger.getLogger(GroupJPAService.class.getName());
+
+	private UserJPAService userJPA = new UserJPAService();
+
 	//The entity manager for this class.
 	private EntityManager entityManager;
 
@@ -27,6 +36,14 @@ public class GroupJPAService{
 		} else {
 			this.entityManager = entityManager;
 		}
+	}
+
+	/**
+	 * Set the userJPAService
+	 * @param userJPA
+	 */
+	public void setUserJPAService(UserJPAService userJPA) {
+		this.userJPA = userJPA;
 	}
 
     /**
@@ -49,13 +66,20 @@ public class GroupJPAService{
      * @param group
      * @return id of the newly added group
      */
-	public int createGroup(Group group) {
-		beginTransaction();
-		entityManager.persist(group);
-		entityManager.flush();
-		int id= group.getId();
-		endTransaction();
-		return id;
+	public int createGroup(Group group) throws GroupNotPersistedException {
+		try {
+			beginTransaction();
+			entityManager.persist(group);
+			entityManager.flush();
+			endTransaction();
+			LOGGER.info("Successfully created a group with group unique key: "+group.getGroupCode());
+			return group.getId();
+		}
+		catch (Exception e) {
+			LOGGER.info("Could not create the group!");
+			throw new GroupNotPersistedException("Could not create the group");
+		}
+
 	}
 
     /**
@@ -63,22 +87,28 @@ public class GroupJPAService{
      * @param id
      * @return a group
      */
-	public Group getGroup(int id) {
-		beginTransaction();
-		return entityManager.find(Group.class, id);
+	public Group getGroup(int id) throws GroupNotFoundException {
+		try {
+			beginTransaction();
+			return entityManager.find(Group.class, id);
+		}
+		catch (Exception e) {
+			LOGGER.info("Could not find a Group with group id: "+id);
+			throw new GroupNotFoundException("Could not find a Group with group id: "+id);
+		}
 	}
 
     /**
      * updateGroup updates a given group with the changed attribute values
      * @param currentGroup
      */
-	public void updateGroup(Group currentGroup) {
+	public Boolean updateGroup(Group currentGroup) throws GroupNotFoundException{
 		beginTransaction();
 		Group group = entityManager.find(Group.class, currentGroup.getId());
 
 		if (group == null) {
-			throw new EntityNotFoundException("Can't find Group for ID "
-					+ currentGroup.getId());
+			LOGGER.info("Could not update group since group not found");
+			throw new GroupNotFoundException("Can't find Group for ID: " + currentGroup.getId());
 		}
 
 		group.setFollowees(currentGroup.getFollowees());
@@ -92,6 +122,15 @@ public class GroupJPAService{
 		group.setMsgs(currentGroup.getMsgs());
 		group.setGroupPassword(currentGroup.getGroupPassword());
 		endTransaction();
+
+		if(group.toString().equals(currentGroup.toString())){
+			LOGGER.info("Updated Group : "+currentGroup.getId());
+			return true;
+		}
+		else {
+			LOGGER.info("Could not update group : "+currentGroup.getId());
+			return false;
+		}
 	}
 
     /**
@@ -100,9 +139,14 @@ public class GroupJPAService{
      * @param id
      * @param user
      */
-	public void addUserToGroup(int id, User user) {
+	public void addUserToGroup(int id, User user) throws GroupNotFoundException{
 		beginTransaction();
 		Group group = entityManager.find(Group.class, id);
+
+		if (group == null) {
+			throw new GroupNotFoundException("Can't find Group for ID: " + id);
+		}
+
 		group.addUser(user);
 		endTransaction();
 	}
@@ -113,12 +157,18 @@ public class GroupJPAService{
      * @param groupName
      * @return a list of groups
      */
-	public List<Group> searchUsingName(String groupName) {
-		String queryString = "SELECT g FROM Group g WHERE g.name = '" + groupName+"'";
+	public List<Group> searchUsingName(String groupName) throws GroupNotFoundException {
+		try {
+			String queryString = "SELECT g FROM Group g WHERE g.name = '" + groupName+"'";
+			beginTransaction();
+			TypedQuery<Group> query = entityManager.createQuery(queryString,Group.class);
+			return query.getResultList();
+		}
+		catch (Exception e) {
+			LOGGER.info("Can't find Group with name: " + groupName);
+			throw new GroupNotFoundException("Can't find Group with name: " + groupName);
+		}
 
-		beginTransaction();
-		TypedQuery<Group> query = entityManager.createQuery(queryString,Group.class);
-		return query.getResultList();
 	}
 
     /**
@@ -126,43 +176,78 @@ public class GroupJPAService{
      * @param groupCode
      * @return a group
      */
-	public Group searchUsingCode(String groupCode) {
-		String queryString = "SELECT g FROM Group g WHERE g.groupCode = '" + groupCode + "'";
-		beginTransaction();
-		Query query = entityManager.createQuery(queryString);
-		return (Group) query.getSingleResult();
+	public Group searchUsingCode(String groupCode) throws GroupNotFoundException {
+		try {
+			String queryString = "SELECT g FROM Group g WHERE g.groupCode = '" + groupCode + "'";
+			beginTransaction();
+			TypedQuery<Group> query = entityManager.createQuery(queryString, Group.class);
+			Group group = query.getSingleResult();
+			endTransaction();
+			return group;
+		}
+		catch (Exception e) {
+			LOGGER.info("Can't find Group with code: " + groupCode);
+			throw new GroupNotFoundException("Can't find Group with code: " + groupCode);
+		}
 	}
-
 
     /**
      * deletes a group from database based on the id of the group
      * @param currentGroup
      */
-	public void deleteGroup(Group currentGroup) {
+	public void deleteGroup(Group currentGroup) throws GroupNotFoundException, GroupNotDeletedException {
 		Group retrievedGroup = searchUsingCode(currentGroup.getGroupCode());
-		beginTransaction();
-		Group group = entityManager.find(Group.class, retrievedGroup.getId());
-		entityManager.remove(group);
-		endTransaction();
+
+		try {
+			beginTransaction();
+			Group group = entityManager.find(Group.class, retrievedGroup.getId());
+			entityManager.remove(group);
+			LOGGER.info("Deleted group with group unique key: "+currentGroup.getGroupCode());
+			endTransaction();
+		}
+		catch(Exception e) {
+			LOGGER.info("Can't delete Group with code: " + currentGroup.getGroupCode());
+			throw new GroupNotDeletedException("Could not delete group with code: "+currentGroup.getGroupCode());
+		}
+
 	}
 
     /**
      * removeUserFromGroup removes the relationship between group and user in the join table
      * @param currentGroup
-     * @param userId
+     * @param username
      * @return 1 for success , 0 for failure
      */
-	public int removeUserFromGroup(Group currentGroup, int userId) {
+	public int removeUserFromGroup(Group currentGroup, String username) throws UserNotFoundException{
 		beginTransaction();
-		int result = entityManager.createNativeQuery("DELETE FROM basegroup_user WHERE user_id="+ userId+
-				" AND group_id="+currentGroup.getId()).executeUpdate();
-		int result2 = entityManager.createNativeQuery("DELETE FROM user_basegroup WHERE user_id="+ userId+
-				" AND groups_id="+currentGroup.getId()).executeUpdate();
-		endTransaction();
-		if(result==1 && result2==1)
-			return 1;
-		else
-			return 0;
+		userJPA.setEntityManager(null);
+		User u = userJPA.search(username);
+		boolean isModerator=false;
+		int userId = u.getId();
+		List<User> moderators = currentGroup.getModerators();
+		for(User user : moderators){
+			if(user.getId()==u.getId()){
+				isModerator=true;
+				break;
+			}
+		}
+		if(!isModerator) {
+			int result = entityManager.createNativeQuery("DELETE FROM basegroup_user WHERE user_id=" + userId +
+					" AND group_id=" + currentGroup.getId()).executeUpdate();
+			int result2 = entityManager.createNativeQuery("DELETE FROM user_basegroup WHERE user_id=" + userId +
+					" AND groups_id=" + currentGroup.getId()).executeUpdate();
+			endTransaction();
+			if(result==1 && result2==1) {
+				ChatLogger.info("Successfully removed User: "+username+" from Group: "+currentGroup.getGroupCode());
+				return 1;
+			}
+			else{
+				ChatLogger.info("Could not remove User: "+username+" from Group: "+currentGroup.getGroupCode());
+				return 0;
+			}
+		}
+		else{
+			throw new IllegalArgumentException("Cannot remove moderator of a group");
+		}
 	}
-
 }
